requirements.txt
requests>=2.31.0
urllib3>=2.0.0
pyodbc>=5.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
tqdm>=4.66.0
speedtest-cli>=2.1.3
schedule>=1.2.0

